+++ dtaianomaly/anomaly_detection/IsolationForest.py
@@ -66,11 +66,11 @@ class IsolationForest(BaseDetector):
         self.window_size = window_size
         self.stride = stride
         self.kwargs = kwargs
-        SklearnIsolationForest(**kwargs)  # Try initialization to check the parameters)
+        SklearnIsolationForest(**kwargs)  # Try initialization to check the parameters
 
     def fit(self, X: np.ndarray, y: Optional[np.ndarray] = None) -> 'IsolationForest':
         """
-        Fit this detector to the given data. Will be called automatically by `decision_function`.
+        Fit this detector to the given data.
 
         Parameters
         ----------