+++ tests/anomaly_detection/test_BaseDetector.py
@@ -2,62 +2,62 @@
 import os
 import pytest
 import numpy as np
-from typing import Optional
 
-from dtaianomaly.anomaly_detection import BaseDetector, load_detector
+from dtaianomaly.anomaly_detection import BaseDetector, load_detector, baselines
 from dtaianomaly import utils
 
 
-class RandomDetector(BaseDetector):
+class InvalidConstantDecisionFunctionForPredictProba(BaseDetector):
 
-    def fit(self, X: np.ndarray, y: Optional[np.ndarray] = None) -> 'RandomDetector':
+    def fit(self, X, y=None):
         return self
 
     def decision_function(self, X: np.ndarray) -> np.ndarray:
-        return np.random.uniform(size=X.shape[0])
-
-
-class AlwaysNormalDetector(BaseDetector):
-
-    def fit(self, X: np.ndarray, y: Optional[np.ndarray] = None) -> 'AlwaysNormalDetector':
-        return self
-
-    def decision_function(self, X: np.ndarray) -> np.ndarray:
-        return np.zeros(shape=X.shape[0])
+        return np.ones(X.shape[0]) * 50
 
 
 class TestBaseDetector:
 
     def test_proba(self):
         data = np.random.standard_normal((50,))
-        detector = RandomDetector()
+        detector = baselines.RandomDetector()
         probas = detector.predict_proba(data)
-        assert 0. <= np.min(probas)
-        assert np.max(probas) <= 1.
+        assert np.all((0.0 <= probas) & (probas <= 1.0))
 
     def test_proba_always_normal(self):
         data = np.random.standard_normal((50,))
-        detector = AlwaysNormalDetector()
+        detector = baselines.AlwaysNormal()
         probas = detector.predict_proba(data)
-        assert 0. == np.min(probas)
-        assert np.max(probas) == 0.0
+        assert np.all(probas == 0.0)
+
+    def test_proba_always_anomalous(self):
+        data = np.random.standard_normal((50,))
+        detector = baselines.AlwaysAnomalous()
+        probas = detector.predict_proba(data)
+        assert np.all(probas == 1.0)
+
+    def test_proba_invalid_constant_decision_function(self):
+        data = np.random.standard_normal((50,))
+        detector = InvalidConstantDecisionFunctionForPredictProba()
+        with pytest.raises(ValueError):
+            detector.predict_proba(data)
 
     def test_proba_invalid(self):
         invalid_data = None
-        detector = RandomDetector()
+        detector = baselines.RandomDetector()
         assert not utils.is_valid_array_like(invalid_data)
         with pytest.raises(ValueError):
             detector.predict_proba(invalid_data)
 
     def test_save_invalid_path(self, tmp_path):
-        detector = RandomDetector()
+        detector = baselines.RandomDetector()
         detector.save(tmp_path / 'some' / 'invalid' / 'directory' / 'testing')
         assert os.path.exists(tmp_path / 'some' / 'invalid' / 'directory')
 
     def test_save_and_load(self, tmp_path):
 
         # Save the detector
-        detector = RandomDetector()
+        detector = baselines.RandomDetector()
         detector.save(tmp_path / 'testing')
 
         # Load the detector
@@ -71,5 +71,5 @@ class TestBaseDetector:
         _ = loaded_detector.predict_proba(data)
 
     def test_str(self):
-        assert str(RandomDetector()) == 'RandomDetector()'
-        assert str(AlwaysNormalDetector()) == 'AlwaysNormalDetector()'
+        assert str(baselines.RandomDetector()) == 'RandomDetector()'
+        assert str(baselines.AlwaysNormal()) == 'AlwaysNormal()'