+++ tests/anomaly_detection/test_MatrixProfileDetector.py
@@ -1,5 +1,6 @@
 
 import pytest
+from sklearn.exceptions import NotFittedError
 from dtaianomaly.anomaly_detection import MatrixProfileDetector
 
 
@@ -49,6 +50,49 @@ class TestMatrixProfileDetector:
         with pytest.raises(ValueError):
             MatrixProfileDetector(window_size=15, k=0)
         MatrixProfileDetector(5, k=2)  # Doesn't raise an error
+
+    def test_initialize_non_bool_novelty(self):
+        with pytest.raises(TypeError):
+            MatrixProfileDetector(window_size=15, novelty=0)
+        MatrixProfileDetector(5, novelty=False)  # Doesn't raise an error
+
+    def test_novelty_univariate(self, univariate_time_series):
+        detector = MatrixProfileDetector(window_size=15, novelty=True)
+        y_pred = detector.fit(univariate_time_series).decision_function(univariate_time_series)
+        assert y_pred.shape == (univariate_time_series.shape[0],)
+
+    def test_novelty_multivariate(self, multivariate_time_series):
+        detector = MatrixProfileDetector(window_size=15, novelty=True)
+        y_pred = detector.fit(multivariate_time_series).decision_function(multivariate_time_series)
+        assert y_pred.shape == (multivariate_time_series.shape[0],)
+
+    def test_not_fitted_no_novelty(self, univariate_time_series):
+        detector = MatrixProfileDetector(window_size=15, novelty=False)
+        detector.decision_function(univariate_time_series)  # No error
+
+    def test_not_fitted_novelty(self, univariate_time_series):
+        detector = MatrixProfileDetector(window_size=15, novelty=True)
+        with pytest.raises(NotFittedError):
+            detector.decision_function(univariate_time_series)
+
+    def test_novelty_different_dimensions(self, univariate_time_series, multivariate_time_series):
+        detector = MatrixProfileDetector(window_size=15, novelty=True)
+        detector.fit(univariate_time_series)
+        with pytest.raises(ValueError):
+            detector.decision_function(multivariate_time_series)
+
+    def test_no_novelty_different_dimension(self, univariate_time_series, multivariate_time_series):
+        detector = MatrixProfileDetector(window_size=15, novelty=False)
+        detector.fit(multivariate_time_series)
+        detector.decision_function(univariate_time_series)  # No error
+
+    def test_novelty_univariate_but_multi_dimensional(self, univariate_time_series):
+        data_2d = univariate_time_series.reshape(-1, 1)
+        assert data_2d.shape == (univariate_time_series.shape[0], 1)
+        assert len(data_2d.shape) == 2
+        detector = MatrixProfileDetector(window_size=15, novelty=True)
+        detector.fit(univariate_time_series).decision_function(data_2d)  # No error
+        detector.fit(data_2d).decision_function(univariate_time_series)  # No error
 
     def test_str(self):
         assert str(MatrixProfileDetector(5)) == "MatrixProfileDetector(window_size=5)"